# web/client/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
    "/api/commands/apply": {
      "post": {
        "summary": "Apply",
        "description": "Apply a plan",
        "operationId": "apply_api_commands_apply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_apply_api_commands_apply_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanApplyStageTracker"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/commands/evaluate": {
      "post": {
        "summary": "Evaluate",
        "description": "Evaluate a model with a default limit of 1000",
        "operationId": "evaluate_api_commands_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EvaluateInput" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/commands/fetchdf": {
      "post": {
        "summary": "Fetchdf",
        "description": "Fetches a dataframe given a sql string",
        "operationId": "fetchdf_api_commands_fetchdf_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FetchdfInput" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/commands/render": {
      "post": {
        "summary": "Render",
        "description": "Renders a model's query, optionally expanding referenced models",
        "operationId": "render_api_commands_render_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RenderInput" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Query" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/commands/test": {
      "get": {
        "summary": "Test",
        "description": "Run one or all model tests",
        "operationId": "test_api_commands_test_get",
        "parameters": [
          {
            "required": false,
            "schema": { "type": "string", "title": "Test" },
            "name": "test",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "Verbose",
              "default": false
            },
            "name": "verbose",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TestResult" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BackfillTask": {
        "properties": {
          "completed": { "type": "integer", "title": "Completed" },
          "total": { "type": "integer", "title": "Total" },
          "view_name": { "type": "string", "title": "View Name" },
          "start": { "type": "integer", "title": "Start" },
          "end": { "type": "integer", "title": "End" }
        },
        "type": "object",
        "required": ["completed", "total", "view_name", "start"],
        "title": "BackfillTask"
      },
      "Body_apply_api_commands_apply_post": {
        "properties": {
          "environment": { "type": "string", "title": "Environment" },
          "plan_dates": { "$ref": "#/components/schemas/PlanDates" },
          "plan_options": { "$ref": "#/components/schemas/PlanOptions" },
          "categories": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SnapshotChangeCategory"
            },
            "type": "object"
          }
        },
        "type": "object",
        "title": "Body_apply_api_commands_apply_post"
      },
...
      "EvaluateInput": {
        "properties": {
          "model": { "type": "string", "title": "Model" },
          "start": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "string", "format": "date-time" },
              { "type": "string" },
              { "type": "integer" },
              { "type": "number" }
            ],
            "title": "Start"
          },
          "end": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "string", "format": "date-time" },
              { "type": "string" },
              { "type": "integer" },
              { "type": "number" }
            ],
            "title": "End"
          },
          "execution_time": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "string", "format": "date-time" },
              { "type": "string" },
              { "type": "integer" },
              { "type": "number" }
            ],
            "title": "Execution Time"
          },
          "limit": { "type": "integer", "title": "Limit", "default": 1000 }
        },
        "type": "object",
        "required": ["model", "start", "end", "execution_time"],
        "title": "EvaluateInput"
      },
      "FetchdfInput": {
        "properties": {
          "sql": { "type": "string", "title": "Sql" },
          "limit": { "type": "integer", "title": "Limit", "default": 1000 }
        },
        "type": "object",
        "required": ["sql"],
        "title": "FetchdfInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanApplyStageTracker": {
        "properties": {
          "start": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "string", "format": "date-time" },
              { "type": "string" },
              { "type": "integer" },
              { "type": "number" }
            ],
            "title": "Start"
          },
          "end": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "string", "format": "date-time" },
              { "type": "string" },
              { "type": "integer" },
              { "type": "number" }
            ],
            "title": "End"
          },
          "meta": { "$ref": "#/components/schemas/TrackableMeta" },
          "environment": { "type": "string", "title": "Environment" },
          "plan_options": { "$ref": "#/components/schemas/PlanOptions" },
          "creation": { "$ref": "#/components/schemas/PlanStageCreation" },
          "restate": { "$ref": "#/components/schemas/PlanStageRestate" },
          "backfill": { "$ref": "#/components/schemas/PlanStageBackfill" },
          "promote": { "$ref": "#/components/schemas/PlanStagePromote" }
        },
        "type": "object",
        "title": "PlanApplyStageTracker"
      },
...
      "PlanDates": {
        "properties": {
          "start": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "string", "format": "date-time" },
              { "type": "string" },
              { "type": "integer" },
              { "type": "number" }
            ],
            "title": "Start"
          },
          "end": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "string", "format": "date-time" },
              { "type": "string" },
              { "type": "integer" },
              { "type": "number" }
            ],
            "title": "End"
          }
        },
        "type": "object",
        "title": "PlanDates"
      },
      "PlanOptions": {
        "properties": {
          "skip_tests": {
            "type": "boolean",
            "title": "Skip Tests",
            "default": false
          },
          "skip_backfill": {
            "type": "boolean",
            "title": "Skip Backfill",
            "default": false
          },
          "no_gaps": {
            "type": "boolean",
            "title": "No Gaps",
            "default": false
          },
          "forward_only": {
            "type": "boolean",
            "title": "Forward Only",
            "default": false
          },
          "no_auto_categorization": {
            "type": "boolean",
            "title": "No Auto Categorization",
            "default": false
          },
          "include_unmodified": {
            "type": "boolean",
            "title": "Include Unmodified",
            "default": false
          },
          "create_from": { "type": "string", "title": "Create From" },
          "restate_models": { "type": "string", "title": "Restate Models" }
        },
        "type": "object",
        "title": "PlanOptions"
      },
...
      "PlanStageBackfill": {
        "properties": {
          "meta": { "$ref": "#/components/schemas/TrackableMeta" },
          "queue": {
            "items": { "type": "string" },
            "type": "array",
            "uniqueItems": true,
            "title": "Queue",
            "default": []
          },
          "tasks": {
            "additionalProperties": {
              "$ref": "#/components/schemas/BackfillTask"
            },
            "type": "object",
            "title": "Tasks",
            "default": {}
          }
        },
        "type": "object",
        "title": "PlanStageBackfill"
      },
...
      "PlanStageCreation": {
        "properties": {
          "meta": { "$ref": "#/components/schemas/TrackableMeta" },
          "total_tasks": { "type": "integer", "title": "Total Tasks" },
          "num_tasks": { "type": "integer", "title": "Num Tasks" }
        },
        "type": "object",
        "required": ["total_tasks", "num_tasks"],
        "title": "PlanStageCreation"
      },
      "PlanStagePromote": {
        "properties": {
          "meta": { "$ref": "#/components/schemas/TrackableMeta" },
          "total_tasks": { "type": "integer", "title": "Total Tasks" },
          "num_tasks": { "type": "integer", "title": "Num Tasks" },
          "target_environment": {
            "type": "string",
            "title": "Target Environment"
          }
        },
        "type": "object",
        "required": ["total_tasks", "num_tasks", "target_environment"],
        "title": "PlanStagePromote"
      },
      "PlanStageRestate": {
        "properties": {
          "meta": { "$ref": "#/components/schemas/TrackableMeta" }
        },
        "type": "object",
        "title": "PlanStageRestate"
      },
...
      "Query": {
        "properties": { "sql": { "type": "string", "title": "Sql" } },
        "type": "object",
        "required": ["sql"],
        "title": "Query"
      },
...
      "RenderInput": {
        "properties": {
          "model": { "type": "string", "title": "Model" },
          "start": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "string", "format": "date-time" },
              { "type": "string" },
              { "type": "integer" },
              { "type": "number" }
            ],
            "title": "Start"
          },
          "end": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "string", "format": "date-time" },
              { "type": "string" },
              { "type": "integer" },
              { "type": "number" }
            ],
            "title": "End"
          },
          "execution_time": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "string", "format": "date-time" },
              { "type": "string" },
              { "type": "integer" },
              { "type": "number" }
            ],
            "title": "Execution Time"
          },
          "expand": {
            "anyOf": [
              { "type": "boolean" },
              { "items": { "type": "string" }, "type": "array" }
            ],
            "title": "Expand",
            "default": false
          },
          "pretty": { "type": "boolean", "title": "Pretty", "default": true },
          "dialect": { "type": "string", "title": "Dialect" }
        },
        "type": "object",
        "required": ["model"],
        "title": "RenderInput"
      },
...
      "SnapshotChangeCategory": {
        "type": "integer",
        "enum": [1, 2, 3, 4, 5],
        "title": "SnapshotChangeCategory",
        "description": "Values are ordered by decreasing severity and that ordering is required.\n\nBREAKING: The change requires that snapshot modified and downstream dependencies be rebuilt\nNON_BREAKING: The change requires that only the snapshot modified be rebuilt\nFORWARD_ONLY: The change requires no rebuilding\nINDIRECT_BREAKING: The change was caused indirectly and is breaking.\nINDIRECT_NON_BREAKING: The change was caused indirectly by a non-breaking change."
      },
...
      "Status": {
        "type": "string",
        "enum": ["init", "success", "fail"],
        "title": "Status",
        "description": "An enumeration of statuses."
      },
...
      "TestErrorOrFailure": {
        "properties": {
          "name": { "type": "string", "title": "Name" },
          "path": { "type": "string", "format": "path", "title": "Path" },
          "tb": { "type": "string", "title": "Tb" }
        },
        "type": "object",
        "required": ["name", "path", "tb"],
        "title": "TestErrorOrFailure"
      },
      "TestResult": {
        "properties": {
          "tests_run": { "type": "integer", "title": "Tests Run" },
          "failures": {
            "items": { "$ref": "#/components/schemas/TestErrorOrFailure" },
            "type": "array",
            "title": "Failures"
          },
          "errors": {
            "items": { "$ref": "#/components/schemas/TestErrorOrFailure" },
            "type": "array",
            "title": "Errors"
          },
          "skipped": {
            "items": { "$ref": "#/components/schemas/TestSkipped" },
            "type": "array",
            "title": "Skipped"
          }
        },
        "type": "object",
        "required": ["tests_run", "failures", "errors", "skipped"],
        "title": "TestResult"
      },
      "TestSkipped": {
        "properties": {
          "name": { "type": "string", "title": "Name" },
          "path": { "type": "string", "format": "path", "title": "Path" },
          "reason": { "type": "string", "title": "Reason" }
        },
        "type": "object",
        "required": ["name", "path", "reason"],
        "title": "TestSkipped"
      },
      "TrackableMeta": {
        "properties": {
          "status": {
            "allOf": [{ "$ref": "#/components/schemas/Status" }],
            "default": "init"
          },
          "start": { "type": "integer", "title": "Start" },
          "end": { "type": "integer", "title": "End" },
          "done": { "type": "boolean", "title": "Done", "default": false }
        },
        "type": "object",
        "title": "TrackableMeta"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}